Add a cart link/icon to the Navbar
This commit also adds a display for the quantity in the cart which is a
variable passed in as a prop to the Navbar component.

The cart icon also functions as a NavLink that takes the user to a cart
route which will eventually display all of the products in the user's
cart.