Implement forms for adding items to cart
A form is rendered for each product in the shop that allows the user to
adjust the product quantity to add to their cart.

A function was created within the App component that calculates the
total quantity of all the products in the cart and displays it in the
Navbar component next to the cart icon.

The bug where the event listeners for the product quantity forms
disappear needs to be fixed because it makes the page refresh if the
user navigates to a different component then re-renders the Shop
component.